guestbost berths map stuff

+++ database/migrations/testing/2022_06_19_190914_create_guest_boat_berths_test_table.php
@@ -19,8 +19,8 @@ class CreateGuestBoatBerthsTestTable extends Migration
             $table->decimal('width', 3, 1)->unsigned()->nullable();
             $table->decimal('length', 3, 1)->unsigned()->nullable();
             $table->decimal('daily_price', 5)->unsigned()->nullable();
-            $table->float('lat', 10, 0)->unsigned()->nullable();
-            $table->float('lng', 10, 0)->nullable();
+            $table->double('lat')->unsigned()->nullable();
+            $table->double('lng')->nullable();
             $table->boolean('enabled')->default(1);
         });
     }
